Move ``uses_shed.py`` to ``galaxy_test.driver``
to break circular dependency between galaxy-test-base and
galaxy-test-driver packages.

Also:
- Use f-string.
