+++ packages/test_base/setup.cfg
@@ -31,8 +31,6 @@ version = 22.5.0.dev0
 [options]
 include_package_data = True
 install_requires =
-    galaxy-app
-    galaxy-data
     galaxy-tool-util
     galaxy-util
     bioblend
