+++ packages/test_driver/setup.cfg
@@ -32,15 +32,14 @@ version = 22.5.0.dev0
 include_package_data = True
 install_requires =
     galaxy-app
+    galaxy-config
     galaxy-data
     galaxy-test-base
     galaxy-tool-util
     galaxy-util
     galaxy-webapps
-    galaxy-web-framework
     nose
     pytest
-    PyYAML
 packages = find:
 python_requires = >=3.7
