+++ test/integration/test_data_manager.py
@@ -8,11 +8,11 @@ from galaxy_test.base.populators import (
     DatasetPopulator,
     skip_if_toolshed_down,
 )
-from galaxy_test.base.uses_shed import (
+from galaxy_test.driver import integration_util
+from galaxy_test.driver.uses_shed import (
     CONDA_AUTO_INSTALL_JOB_TIMEOUT,
     UsesShed,
 )
-from galaxy_test.driver import integration_util
 
 FETCH_TOOL_ID = "toolshed.g2.bx.psu.edu/repos/devteam/data_manager_fetch_genome_dbkeys_all_fasta/data_manager_fetch_genome_all_fasta_dbkey/0.0.3"
 FETCH_GENOME_DBKEYS_ALL_FASTA_INPUT = {
