+++ test/integration_selenium/test_workflow_repository_tool_update.py
@@ -1,4 +1,4 @@
-from galaxy_test.base.uses_shed import UsesShed
+from galaxy_test.driver.uses_shed import UsesShed
 from .framework import (
     selenium_test,
     SeleniumIntegrationTestCase,
